parseUnitPrefix now takes a double

+++ include/GenericToolbox.h
@@ -118,7 +118,7 @@ namespace GenericToolbox{
   inline void indentInputString(std::string& inputStr_, unsigned int indentCount_, const std::string& indentChar = " ");
 
   // -- Parsing
-  inline std::string parseUnitPrefix(long val_);
+  inline std::string parseUnitPrefix(double val_);
   inline std::string parseSizeUnits(unsigned int sizeInBytes_);
   inline std::string formatString( const std::string& strToFormat_ ); // overrider: make sure this is the one used when no extra args are provided.
   template<typename ... Args> inline std::string formatString(const std::string& strToFormat_, Args ... args );
